requirements.txt
Django==5.0.0
pymongo==4.5.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
dnspython==2.4.2
Pillow==10.4.0
reportlab==4.0.7